Security update.  Added Orders (shipments) and Items, fixed some stuff in schematics which still doesn't turn on valves properly for dual outputs (see the TankField schematic) Still in the process of handling integration of Orders w/the initiation of transfers.  Currently there are 4 scheduled transfers for refreshing the crude inventory.  Anyway, i need to add handling the reception of docks and trains and tank trucks ...

+++ oms/src/main/java/us/avn/oms/mapper/TransferMapper.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2018 A. E. Van Ness
+ * Copyright (C) 2018 Laboratorio de Lobo Azul
  *  
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -51,6 +51,6 @@ public interface TransferMapper {
 	
 	Integer updateTransfer( Transfer x );
 
-	Integer insertTransfer( Transfer x );
+	Long insertTransfer( Transfer x );
 
 }
